+++ src/domain/auth/type/AuthType.ts
