+++ src/domain/layout/component/sidebar/SideLogo.tsx
