+++ src/domain/post/type/PostType.ts
